### 2019-06-08 #1495 @whedon
[PRE REVIEW]: SmartEDA: An R Package for Automated Exploratory Data Analysis
**Submitting author:** @sayanddude (<a href="http://orcid.org/0000-0002-6477-0376">Sayan Putatunda</a>)
**Repository:** <a href="https://github.com/daya6489/SmartEDA" target ="_blank">https://github.com/daya6489/SmartEDA</a>
**Version:** v0.3.1
