### 2019-06-08 #1495 @sayanddude
> PDF failed to compile for issue #1495 with the following error:
>
> Error producing PDF.
> ! Undefined control sequence.
> l.444 the ISLR package \citep
>
> Looks like we failed to compile the PDF

please compile the pdf again. I have corrected the issue.
