+++ README.Rmd
@@ -1,0 +1,467 @@
+---
+output: github_document
+---
+  
+```{r setup, echo=FALSE}
+library(rmarkdown)
+library(SmartEDA)
+library(knitr)
+library(ISLR)
+library(scales)
+library(gridExtra)
+library(ggplot2)
+
+knitr::opts_chunk$set(
+	eval = FALSE,
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-"
+)
+	
+```
+
+
+# SmartEDA [![CRAN status](https://www.r-pkg.org/badges/version/SmartEDA)](https://cran.r-project.org/package=SmartEDA)
+
+[![Downloads](http://cranlogs.r-pkg.org/badges/SmartEDA)](https://cran.r-project.org/package=SmartEDA)
+[![Total Downloads](http://cranlogs.r-pkg.org/badges/grand-total/SmartEDA)](https://cran.r-project.org/package=SmartEDA)
+
+---
+
+# Background
+In a quality statistical data analysis the initial step has to be exploratory. Exploratory data analysis begins with the univariate exploratory analyis - examining the variable one at a time. Next comes bivariate analysis followed by multivariate analyis. SmartEDA package helps in getting the complete exploratory data analysis just by running the function instead of writing lengthy r code.
+
+# Functionalities of SmartEDA
+
+The SmartEDA R package has four unique functionalities as
+
+* Descriptive statistics
+* Data visualisation
+* Custom table
+* HTML EDA report
+
+![SmartEDA](https://github.com/daya6489/SmartEDA/blob/master/man/figures/smarteda_funtions.PNG)
+
+
+# Journal of Open Source Software Article
+An article describing SmartEDA pacakge for exploratory data analysis approach has been published in [arxiv](https://arxiv.org/pdf/1903.04754.pdf) and currently it is under review at The Journal of Open Source Software. Please cite the paper if you use SmartEDA in your work!
+
+# Installation
+
+The package can be installed directly from CRAN.
+
+```R
+install.packages("SmartEDA")
+```
+
+You can install the latest development verion of the [SmartEDA](https://github.com/daya6489/SmartEDA) from github with:
+	
+```R
+install.packages("devtools")
+devtools::install_github("daya6489/SmartEDA",ref = "develop")
+```
+
+# Example
+
+## Data
+
+In this vignette, we will be using a simulated data set containing sales of child car seats at 400 different stores. 
+
+Data Source [ISLR package](https://www.rdocumentation.org/packages/ISLR/versions/1.2/topics/Carseats).
+
+Install the package "ISLR" to get the example data set.
+
+```{r eda-c3-r, warning=FALSE,eval=F}
+install.packages("ISLR")
+library("ISLR")
+install.packages("SmartEDA")
+library("SmartEDA")
+## Load sample dataset from ISLR pacakge
+Carseats= ISLR::Carseats
+```
+
+### Overview of the data
+Understanding the dimensions of the dataset, variable names, overall missing summary and data types of each variables
+
+```{r od_1,warning=FALSE,eval=F,include=T}
+# Overview of the data - Type = 1
[…]
