+++ DESCRIPTION
@@ -1,13 +1,13 @@
 Package: SmartEDA
 Type: Package
 Title: Summarize and Explore the Data
-Version: 0.3.1
+Version: 0.3.2
 Author: Dayanand Ubrangala; Kiran R; Ravi Prasad Kondapalli; Sayan Putatunda
 Maintainer: Dayanand Ubrangala<daya6489@gmail.com>
 Depends: R (>= 3.3.0)
 Imports: ggplot2(>= 2.2.1),sampling(>= 2.8),scales(>= 0.5.0),rmarkdown(>= 1.9),ISLR(>= 1.2),data.table(>= 1.10.4-3),gridExtra(>= 2.3),GGally(>= 1.4.0)
 Description: Exploratory analysis on any input data describing the structure and the relationships present in the data. The package automatically select the variable and does related descriptive statistics. Analyzing information value, weight of evidence, custom tables, summary statistics, graphical techniques will be performed for both numeric and categorical predictors.
-License: MIT + file LICENSE
+License: GPL-3
 Suggests: psych, Hmisc, smbinning,testthat,knitr,covr
 Encoding: UTF-8
 LazyData: true
