### 2019-06-13 faf03bc
SmartEDA page updated
