### 2019-06-13 7a44a7d
JOSS paper manuscript
