+++ paper/paper.md
@@ -1,0 +1,162 @@
+---
+title: "SmartEDA: An R Package for Automated Exploratory Data Analysis"
+tags:
+  - R
+  - Exploratory Data Analysis
+  - Data Mining
+authors:
+  - name: Sayan Putatunda
+    orcid: 0000-0002-6477-0376
+    affiliation: 1
+  - name: Dayananda Ubrangala
+    affiliation: 1
+  - name: Kiran R
+    affiliation: 1
+  - name: Ravi Kondapalli
+    affiliation: 1
+affiliations:
+  - name: VMware Software India Pvt ltd.
+    index: 1
+date: 04 June 2019
+bibliography: paper.bib
+---
+
+
+# Introduction: Exploratory Data Analysis
+
+Nowadays, we see applications of Data Science almost everywhere. Some of the well highlighted aspects of data science are the various statistical and machine learning techniques applied for solving a problem. However, any data science activity starts with an Exploratory Data Analysis (EDA). The term "Exploratory Data Analysis" was coined by [@eda:1]. EDA can be defined as the art and science of performing initial investigation on the data by means of statistical and visualization techniques that can bring out the important aspects in the data that can be used for further analysis [@eda:1]. EDA puts an emphasis on hypothesis generation and pattern recognition from the raw data [@eda:8]. There have been many studies conducted on EDA reported in the Statistics literature.Some of earliest work done on Exploratory Data Analysis (EDA) including coining the term and defining some of the basic EDA techniques was done by [@eda:1]. However, many researchers have formulated different definitions of EDA over the years. One of the widely accepted definition is that "Exploratory data analysis isolates patterns and features of the data and reveals these forcefully to the analyst" [@eda:12].
+
+@eda:4 introduced EDA in the context of data mining and resampling with focus on pattern recognition, cluster detection and variable selection. Over the years, EDA has been used various applications across different domains such as geoscience research [@eda:7], auditing [@eda:8], game-based assessments [@eda:9], clinical study groups [@eda:10] and more.
+
+![The CRISP-DM Process Model for Data Mining @crisp:1.](paper_figures/Fig1.png){width="5in"}
+
+EDA is a very important component of the Data Mining process as per the industry standard CRSP-DM framework. The CRISP-DM stands for "CRoss Industry Standard Process for Data Mining" [@crisp:1]. Data mining is a creative process that required a different set of skills and knowledge. However, earlier there was a lack of any standard framework for Data mining projects, which esured that the sucess/failure of a data mining project is highly dependent on the skill-set of a particular individual or a team that is executing the project. To address this need, @crisp:1 proposed the CRISP-DM process mopdel that  is a framework for executing any data mining project. The CRISP-DM framework is independent of the tools used and the industry sector. 
+
+Figure 1 shows the different components (viz. Business understanding, Data understanding, Data preparation, Modeling, Evaluation and Deployment) of the CRISP-DM process model for data mining. It is a cyclic process where there is a feedback loop between some components. We can see that the "Data Understanding" is a very important component which affects the "Business Understanding" as well. EDA helps in Data understanding and thus directly impacts the quality and success of a data mining project.
+
[…]
