+++ paper/paper.bib
@@ -1,249 +1,0 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% http://bibdesk.sourceforge.net/
-
-%% Created for Sayan Putatunda at 2019-03-07 16:34:31 +0530 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-
-@manual{pack:8,
-	Author = {Hadley Wickham},
-	Date-Added = {2019-03-07 16:31:06 +0530},
-	Date-Modified = {2019-03-07 16:34:28 +0530},
-	Title = {Flights that Departed NYC in 2013},
-	Url = {{http://github.com/hadley/nycflights13}},
-	Year = {2018}}
-
-@manual{pack:7,
-	Author = {Gareth James and Daniela Witten and Trevor Hastie and Rob Tibshirani},
-	Date-Added = {2019-03-04 13:00:46 +0530},
-	Date-Modified = {2019-03-04 13:01:32 +0530},
-	Title = {{ISLR: Data for an Introduction to Statistical Learning with Applications in R}},
-	Url = {{https://CRAN.R-project.org/package=ISLR}},
-	Year = {2017}}
-
-@article{eda:12,
-	Author = {Hoaglin, D. and Mosteller, F. and Tukey, J.W.},
-	Date-Added = {2019-01-09 19:50:32 +0530},
-	Date-Modified = {2019-01-09 19:51:26 +0530},
-	Journal = {Wiley Series in probability and mathematical statistics, New-York},
-	Title = {Understanding Robust and Exploratory Data Analysis},
-	Year = {1983}}
-
-@manual{eda:11,
-	Author = {Jaggi, S.},
-	Date-Added = {2019-01-09 15:06:33 +0530},
-	Date-Modified = {2019-01-09 17:03:08 +0530},
-	Month = {February},
-	Organization = {Indian Agricultural Statistics Research Institute},
-	Title = {Descriptive Statistics and Exploratory Data Analysis},
-	Url = {http://www.iasri.res.in/ebook/EB_SMAR/e- book_pdf%20files/Manual%20II/1-Descriptive%20Statistics.pdf},
-	Year = {2013},
-	Bdsk-Url-1 = {http://www.iasri.res.in/ebook/EB_SMAR/e-%20book_pdf%20files/Manual%20II/1-Descriptive%20Statistics.pdf}}
-
-@inproceedings{crisp:1,
-	Author = {R{\"u}diger Wirth},
-	Booktitle = {{Proceedings of the Fourth International Conference on the Practical Application of Knowledge Discovery and Data Mining}},
-	Date-Added = {2019-01-08 12:47:40 +0530},
-	Date-Modified = {2019-02-20 14:35:41 +0530},
-	Pages = {29--39},
-	Title = {{CRISP-DM: Towards a standard process model for data mining}},
-	Year = {2000}}
-
-@article{eda:10,
-	Author = {{Konopka et al.}},
-	Date-Added = {2019-01-03 15:59:25 +0530},
-	Date-Modified = {2019-03-05 14:19:54 +0530},
-	Journal = {{PLoS ONE}},
-	Number = {8},
-	Title = {{Exploratory data analysis of a clinical study group: Development of a procedure for exploring multidimensional data}},
-	Volume = {13},
-	Year = {2018}}
-
-@inbook{eda:9,
-	Address = {Cham},
-	Author = {{DiCerbo et al.}},
-	Chapter = {{An Application of Exploratory Data Analysis in the Development of Game-Based Assessments}},
[…]
