+++ R/fn_exp_categorical_viz.R
@@ -1,235 +1,207 @@
 #' Distributions of categorical variables
 #'
+#'This function automatically scans through each variable and creates bar plot for categorical variable.
 #'
-#' @description This function automatically scans through each variable and creates bar plot for categorical variable.
-#' @usage ExpCatViz(data,target=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NULL,
-#' Flip=F,sample=NULL,rdata=FALSE,value=NULL,gtitle=NULL,theme="Default")
-##' @param data dataframe or matrix
-##' @param target target variable. This is not a mandatory field.
-##' @param fname output file name. Output will be generated in PDF format
-##' @param clim maximum categories to be considered to include in bar graphs.
-##' @param col define the colors to fill the bars, default it will take sample colours.
-##' @param margin index, 1 for row based proportions and 2 for column based proportions
-##' @param Page output pattern. if Page=c(3,2), It will generate 6 plots with 3 rows and 2 columns
-##' @param Flip default vertical bars. It will be used to flip the axis vertical to horizontal
-##' @param sample random selection of categorical variable
-##' @param rdata to plot bar graph for frequency/aggregated table
-##' @param value value coloumn name. This is mandatory if 'rdata' is TRUE
-##' @param gtitle graph title
-##' @param theme extra themes, geoms, and scales for 'ggplot2' (use theme from ggthemes package)
-##' @seealso \code{\link[ggplot2:geom_bar]{geom_bar}}
-##' @return This function returns collated graphs in grid format in PDF or JPEG format. All the files will be stored in the working directory
-##'
-##'Bar graph - for raw data(this function will dynamically pick all the categorical variable and plot the bar chart)
-##'
-##'Bar graph - aggregated data
-##'
-##'Stacked Bar graph by target variable
-##'
-##' @examples
+#' @param data dataframe or matrix
+#' @param target target variable. This is not a mandatory field
+#' @param fname output file name. Output will be generated in PDF format
+#' @param clim maximum categories to be considered to include in bar graphs
+#' @param col define the colors to fill the bars, default it will take sample colours
+#' @param margin index, 1 for row based proportions and 2 for column based proportions
+#' @param Page output pattern. if Page=c(3,2), It will generate 6 plots with 3 rows and 2 columns
+#' @param Flip default vertical bars. It will be used to flip the axis vertical to horizontal
+#' @param sample random selection of categorical variable
+#' @param rdata to plot bar graph for frequency/aggregated table
+#' @param value value coloumn name. This is mandatory if 'rdata' is TRUE
+#' @param gtitle graph title
+#' @param theme adding extra themes, geoms, and scales for 'ggplot2' (eg: themes options from ggthemes package)
+#' @seealso
+#' \code{\link[ggplot2:geom_bar]{geom_bar}}
+#'
+#' @return This function returns collated graphs in grid format in PDF or JPEG format. All the files will be stored in the working directory
+#'
+#' \itemize{
[…]
