### 2019-06-18 #1509 @whedon
[REVIEW]: SmartEDA: An R Package for Automated Exploratory Data Analysis
**Submitting author:** @sayanddude (<a href="http://orcid.org/0000-0002-6477-0376">Sayan Putatunda</a>)
**Repository:** <a href="https://github.com/daya6489/SmartEDA" target ="_blank">https://github.com/daya6489/SmartEDA</a>
**Version:** 0.3.2
