### 2019-06-19 #1509 @terrytangyuan
I saw that you submitted for review to JSS already https://arxiv.org/pdf/1903.04754.pdf. Is this submission to JOSS still necessary?
