### 2019-06-20 #1509 @sayanddude
@terrytangyuan Hi! We initially submitted the paper (given that it's a CRAN package) to JSS a couple of months back but it didn't work out there. We assure you that currently the paper is not under consideration for submission at any journal or conference. We are hoping that this paper goes through the rigorous review and gets published at JOSS!
