### 2019-07-11 #1509 @terrytangyuan
@sayanddude I see. There are many other tools for exploratory data analysis in R. What makes this package useful/different?
