### 2019-07-12 #1509 @sayanddude
> @sayanddude I see. There are many other tools for (automated) exploratory data analysis in R. What makes this package useful/different?

@terrytangyuan Hi! We have added a new section “Comparison with other R Packages” in the updated version of the paper pdf generated above. In this section we have given a snapshot of various capabilities of SmartEDA vs. some of the competing R packages (such as dlookr, explorer, DataExplorer, etc.) and have highlighted its advantages. Figure 9 in the paper gives a snapshot of the comparison and shows how it’s better than most of the available R packages for automated exploratory data analysis (Please find attached the figure below).

To summarize, some of the key benefits of SmartEDA are:
•	No need remember the different R package names as SmartEDA has most of the exploratory function and dependencies
•	No need to write lengthy R scripts. SmartEDA does the exploratory in one line R script
•	It cuts down time for exploratory data analysis
•	SmartEDA has the extension of data.table to build customized summary statistics and cross tables
•	SmartEDA function can generate 100’s of ggplot (like scatter, bar, stacked bar, boxplot, density, qqplot, co-ordinate plots) at a time with customized theme using ggthemes package options

Also, SmartEDA finds mention in the study conducted by Staniak and Biecek (2018) where they reviewed the landscape of R packages for automated Exploratory analysis. Some of the distinguishing features of SmartEDA pointed out by the authors when comparing it with other R packages are:
•	The SmartEDA package  reports skewness and displays QQ plots against normal distribution
•	SmartEDA package provides a method of visualizing multivariate relationships - parallel coordinate plot.
•	SmartEDA give a reasonable insight into variables distributions and simple relationships.
•	Parallel Co-ordinates Plots (PCP) in SmartEDA is unique and is very well done.
This paper by Staniak and Biecek (2019)  is available in ArXiv- https://arxiv.org/pdf/1904.02101.pdf

Reference:
Mateusz Staniak and  Przemyslaw Biecek (2019), “The Landscape of R Packages for Automated Exploratory Data Analysis”, arXiv:1904.02101 [stat.CO]
