### 2019-07-12 #1509 @sayanddude
> 👉 Check article proof 📄 👈

@nhejazi @terrytangyuan Requesting the reviewers to kindly consider the latest version of the pdf. We have added a section on “Comparison with other R Packages” and corrected some formatting issues in the earlier version.
