### 2019-07-12 #1509 @terrytangyuan
@sayanddude Thanks. The table looks great. Some feedback:
* I see lots of long functions with >150 lines of code and the code style is not consistent (I suggest running a `lintr` check). Similarly, in the `roxygen` docs, the style is not consistent. I see both `##'` and `#'`. The indentation levels and the `roxygen` syntax are sometimes incorrect. Please double check.
* There are no tests.
* Have you tried running this package on large datasets?
* Please make the paper more concise - I see many pages where each page only has one giant picture.
