### 2019-07-12 #1509 @sayanddude
@terrytangyuan Thanks for the feedback! I will work on your comments and will get back to you with the updated paper and required code changes as soon as possible.
