### 2019-08-06 #1509 @sayanddude
@terrytangyuan  We have worked on addressing all of your comments. Please find below our response to each action items:

•	_I see lots of long functions with >150 lines of code and the code style is not consistent (I suggest running a lintr check). Similarly, in the roxygen docs, the style is not consistent. I see both ##' and #'. The indentation levels and the roxygen syntax are sometimes incorrect. Please double check._
-	Thanks for your comment! We have updated the Github repository and we have ensured the length of almost all the functions are below 150 lines of code. We have also run a lintr check and have corrected all the issues related to style inconsistencies, indentation issues and incorrect syntaxes.

•	_There isn't any unit test for the package._
-	Thanks a lot for the comment! We have now implemented the unit test of the package (available at https://github.com/daya6489/SmartEDA).

•	_Have you tried running this package on large datasets?._
-	Yes, the package works well on large datasets. Recently, we applied the SmartEDA package on the Microsoft malware prediction data in Kaggel (available at- https://www.kaggle.com/ajithvallabai/microsoft-malware-prediction/data ). This dataset was considerably large i.e. it had 8900000 rows and 82 columns. The SmartEDA package worked seamlessly on this dataset.

•	_Please make the paper more concise - I see many pages where each page only has one giant picture._
-	Thanks for the comment! We have considerably reduced the size of the paper (please refer to the updated version of the paper i.e. **10.21105.joss.01509.pdf** ) by removing a few verbose content and also by consolidating most of the images into a single figure i.e. Figure 2.

Thanks and Regards,
Sayan
