### 2019-08-06 #1509 @sayanddude
@labarba Hi! As discussed, we have completed working on the comments and have updated the Github code repository along with the paper.
