### 2019-08-06 #1509 @labarba
The handling editor, @mgymrek, will take it from here.
