### 2019-08-07 a8e2cf7
added colour function
