### 2019-08-07 4ad058a
updated test function
