### 2019-08-07 97b2638
Updated contribution guidelines
