+++ Paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for Sayan Putatunda at 2019-07-12 14:24:02 +0530 
+%% Created for Sayan Putatunda at 2019-08-19 13:03:01 +0530 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -20,7 +20,8 @@
 @manual{pack:7,
 	Author = {Gareth James and Daniela Witten and Trevor Hastie and Rob Tibshirani},
 	Date-Added = {2019-03-04 13:00:46 +0530},
-	Date-Modified = {2019-03-04 13:01:32 +0530},
+	Date-Modified = {2019-08-16 15:21:18 +0530},
+	Doi = {https://doi.org/10.1007/978-1-4614-7138-7_1},
 	Title = {{ISLR: Data for an Introduction to Statistical Learning with Applications in R}},
 	Url = {{https://CRAN.R-project.org/package=ISLR}},
 	Year = {2017},
@@ -45,56 +46,43 @@
 	Year = {2013},
 	Bdsk-Url-1 = {http://www.iasri.res.in/ebook/EB_SMAR/e-%20book_pdf%20files/Manual%20II/1-Descriptive%20Statistics.pdf}}
 
-@inproceedings{crisp:1,
-	Author = {R{\"u}diger Wirth},
-	Booktitle = {{Proceedings of the Fourth International Conference on the Practical Application of Knowledge Discovery and Data Mining}},
-	Date-Added = {2019-01-08 12:47:40 +0530},
-	Date-Modified = {2019-02-20 14:35:41 +0530},
-	Pages = {29--39},
-	Title = {{CRISP-DM: Towards a standard process model for data mining}},
-	Year = {2000}}
-
 @article{eda:10,
 	Author = {{Konopka et al.}},
 	Date-Added = {2019-01-03 15:59:25 +0530},
-	Date-Modified = {2019-03-05 14:19:54 +0530},
+	Date-Modified = {2019-08-16 14:41:03 +0530},
+	Doi = {https://doi.org/10.1371/journal.pone.0201950},
 	Journal = {{PLoS ONE}},
 	Number = {8},
 	Title = {{Exploratory data analysis of a clinical study group: Development of a procedure for exploring multidimensional data}},
 	Volume = {13},
-	Year = {2018}}
+	Year = {2018},
+	Bdsk-Url-1 = {https://doi.org/10.1371/journal.pone.0201950}}
 
 @inbook{eda:9,
 	Address = {Cham},
 	Author = {{DiCerbo et al.}},
 	Chapter = {{An Application of Exploratory Data Analysis in the Development of Game-Based Assessments}},
 	Date-Added = {2019-01-03 15:51:47 +0530},
-	Date-Modified = {2019-03-05 14:19:33 +0530},
+	Date-Modified = {2019-08-16 15:11:47 +0530},
+	Doi = {10.1007/978-3-319-05834-4},
 	Editor = {Loh, C. and Sheng, Y. and Ifenthaler, D.},
 	Publisher = {Springer},
 	Title = {{Serious Games Analytics. Advances in Game-Based Learning}},
-	Year = {2015}}
-
-@phdthesis{eda:8,
-	Address = {{Newark, New Jersey}},
-	Author = {Qi Liu},
-	Date-Added = {2019-01-03 15:40:32 +0530},
-	Date-Modified = {2019-01-03 15:42:23 +0530},
-	Month = {October},
-	School = {{Newark Rutgers, The State University of New Jersey}},
[…]
