+++ CONTRIBUTING.md
@@ -1,0 +1,48 @@
+# Contribution Guidelines
+
+We, the authors of the SmartEDA R package, use the same guide as is used for contributing to the development of R packages such as `data.table` and `biotmle`. This document is simply a formal re-statement of that fact.
+
+All developers are very much welcome to contribute to code, enhance, provide documentation, 
+and perfomr testing. We look forward to more suggestions from your end.
+
+SmartEDA is a new R package. If you're new to coding and submitting pull requests (PRs), 
+we encourage you to try. You can submit some recommendations in the issues or start coding a 
+function or adding some parameters or fixing a bug or just rewriting a bit so that it is clear, 
+well documented and understandable. Please do not hesitate to submit your changes. We will work 
+together to integrate them.
+
+-----
+
+## Code
+
+Authors name would be added for their code contribution in the description file. Please document 
+and comment your code. This will ensure that the entire code is clear and understandable.
+
+  - Before submitting a change, please use styler to nicely format your code.
+  - Before you start to code, make sure you're on the "dev" branch
+  - Before submitting the final code, please run a lintr check
+  - Please Use '<-' for assignment (not =)
+
+-----
+
+## Issues
+
+Please read and follow all the instructions 
+
+  - check NEWS first and search existing Issues.
+  - Report One issue for one purpose. Don't report more than one bug or request several features in the 
+  same issue.
+  - Spend a little bit of time ensuring that your **code** is easy for others to read
+
+-----
+
+## Pull requests (PRs)
+
+To contribute a change to `SmartEDA`
+
+  - Create a branch in git and make your changes.
+  - Discuss the pull request.
+  - Push branch to GitHub and issue pull request.
+  - Every new feature or bug fix must have one or more new tests.
+
+-----

[…]
