### 2019-08-08 #1509 @mgymrek
@sayanddude thanks for making these changes

@nhejazi, @terrytangyuan can you now go over the revision? If your comments have been sufficiently addressed please finish filling out the checklist
