### 2019-08-15 #1509 @nhejazi
Weighing in with a few points that I think ought to be addressed as we continue the review for the `SmartEDA` package.

* Major (cannot be checked off yet)
  * __References__: I believe some of the references are missing DOIs. As an example, the paper "Exploratory data analysis of a clinical study group: Development of a procedure for exploring multidimensional data" has a DOI ( https://doi.org/10.1371/journal.pone.0201950
), available at https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0201950. Please double-check all references for DOIs and add them where missing.
  * __Community Guidelines__: While the package repository contains a `CODE_OF_CONDUCT.md` file, a "Contributions" section ought to be added to the package `README.md` (and would translate to the `pkgdown` site). It can be something simple that points to the existing `CODE_OF_CONDUCT.md` document (e.g., from a package I wrote https://github.com/nhejazi/biotmle#contributions). Similarly, a statement in the `README.md` about issues would be a welcome addition https://github.com/nhejazi/biotmle#issues.

* Minor (things that will not hold up the review):
  * __Unit tests__: it looks like you've added a number of unit tests --- thank you for addressing this concern brought up by another reviewer. From a quick look at some of these tests, they seems as though they could be better tightened. For example, there's a unit test of a plotting function that simply checks whether the produced object is, in fact, a `ggplot` object. While Visualizations can be particularly difficult to test, there are also some tests that check only that a returned summary statistic matches the values produced by a call to the corresponding base function (e.g., a call to `mean`). Over time, and as your package attracts more users, I would encourage you to write more tests.
  * __JOSS paper__: From a quick read, there are some notational/grammatical/stylistic improvements that could be made. A few examples: (1) In the first paragraph, there are two sentences where an inline reference to Tukey is intended but not used ("was coined by (J.W. Tukey)"; "was done by (J.W. Tukey)"). Note that the correct markdown syntax for such citations would simply be "@eda:1" (rather than the currently used "[@eda:1]"). (2) At the end of the introduction, you refer to R as a statistical computing package, whereas it might better be called a programming environment or language; moreover, you cite Venables & Ripley where you should cite the R core development team. (3) There are some grammatical mistakes in the paper, please re-read it to catch these --- it would be best if they were absent from the final version that gets published.
