### 2019-08-19 #1509 @sayanddude
@nhejazi We have worked on addressing all of your comments. Please find below our response to each action items:

**Major (cannot be checked off yet)**

- **_References: I believe some of the references are missing DOIs. As ..........._**
Thanks for your comment! We have added the DOI for almost all the articles for which DOIs are available in the updated version of the paper. Only for the cited R packages and for the Exploratory Data Analysis book by Tukey (1977), there's no DOI available.

- **_Community Guidelines: While the package repository contains a CODE_OF_CONDUCT.md file, a "Contributions" section .........................._**
Thanks for pointing it out! We have added an "Issues" section and a "Contributions" section in the README.md file of the package (please refer to https://github.com/openjournals/joss-reviews/issues/1509). We have also added a "Contribution Guidelines" as suggested by you.

**Minor (things that will not hold up the review):**

- _**Unit tests: it looks like  .............**_
Thanks for your valuable comment! Yes, we agree that more tests can be written for testing the summary statistics. And yes, as more users start using our package we will keep writing more unit tests. In fact, we plan to write some more unit test in the next release version of the package.

- **_JOSS paper: From a quick read, there are some ................._**
Thanks for your comment! We have corrected the required markdown syntax (i.e. @eda:1). We have also cited the R core development team for "R" and have called R as a "programming environment" in text (pg. 1, paragraph 3, Introduction section) instead of "a statistical computing package" that was mentioned in the earlier version of the paper. We have also corrected some of the grammatical mistakes in the paper and have also tools such as "Grammarly" to run a check of the entire manuscript for any grammatical errors.
