### 2019-09-01 #1509 @mgymrek
Thanks!

@sayanddude see next steps below.

Some minor comments on the paper:
* "EDA can be categorized into Descriptive statistical techniques and graphical techniques": uncapitalize descriptive
* You can reference just Figure 2 rather than spelling out (a) through (f)
* There is a reference to a Figure 9 that I believe should be to Figure 3 instead

After fixing those typos, can you please make a zenodo archive, being sure the title and author list match those on the paper, and report the DOI here?
