### 2019-09-01 #1509 @terrytangyuan
@sayanddude Thanks for addressing the comments. The paper looks good to me now and I recommend for publication.
