### 2019-09-02 #1509 @sayanddude
> Thanks!
>
> @sayanddude see next steps below.
>
> Some minor comments on the paper:
>
> * "EDA can be categorized into Descriptive statistical techniques and graphical techniques": uncapitalize descriptive
> * You can reference just Figure 2 rather than spelling out (a) through (f)
> * There is a reference to a Figure 9 that I believe should be to Figure 3 instead
>
> After fixing those typos, can you please make a zenodo archive, being sure the title and author list match those on the paper, and report the DOI here?

@mgymrek Thanks for the feedback! I have corrected all the required typos in the updated version of the pdf.
Also, I have created the zenodo archive for the SmartEDA package (version 0.3.2) with DOI: 10.5281/zenodo.3383824 and the URL is: https://doi.org/10.5281/zenodo.3383824.

I have ensured that the title and the author names are same as the ones mentioned in the paper. Please let me know if I have missed out on anything. Thank you so much!
