### 2019-09-04 #1509 @mgymrek
Thanks @sayanddude!
@openjournals/joss-eics this paper is ready to be accepted!
