### 2019-09-28 #1509 @terrytangyuan
@arfon Should we close this issue now that the paper has been deposited?
